+++ 2_Back_End/back_end/src/main/java/com/app/dao/OrderDetailRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.query.Param;
 
 import com.app.pojo.OrderDetail;
 
-public interface OrderDetailRepository extends JpaRepository<OrderDetail, Long> {
+public interface OrderDetailRepository extends JpaRepository<OrderDetail, Integer> {
 	@Query("Select od from OrderDetail od where od.currentOrder.id=:id")
-	List<OrderDetail> findAllByOrderId(@Param("id") Long id);
+	List<OrderDetail> findAllByOrderId(@Param("id") Integer id);
 }
