+++ 2_Back_End/back_end/src/main/java/com/app/dto/OrderPlaceDto.java
@@ -2,7 +2,7 @@ package com.app.dto;
 
 public class OrderPlaceDto {
 	private String paymentMode;
-	private long addressId;
+	private int addressId;
 
 	public OrderPlaceDto() {
 	}
@@ -21,7 +21,7 @@ public class OrderPlaceDto {
 		this.paymentMode = paymentMode;
 	}
 
-	public long getAddressId() {
+	public int getAddressId() {
 		return addressId;
 	}
